[mlir] Optimize Operation field count memory usage
Various components within an operation (e.g. regions, results,
successors) have a small number of elements in the overwhelmingly
common case. This commit refactors Operation such that in the
common case, the counts of these fields are stored within a u8,
and in the extreme case the counts are stored in a trailing
object that allows for a full u32 number of elements. This allows
for dropping the size of Operation by 8 bytes (in an x64 we are
now down to 56 bytes).

Differential Revision: https://reviews.llvm.org/D113156