[mlir-lsp-server] Optimize definition lookups from source locations
We currently use a brute force lookup when trying to match a source location to an MLIR definition. This was fine when initially building up the server, but is horrendously slow for large modules. This revision refactors to use an interval map index as the method of mapping a source position to a definition.