[mlir][Pass] Deprecate FunctionPass in favor of a more general SymbolDefinitionPass
This revision generalizes the functionality currently present in FunctionPass, i.e.
running a pass on symbol definitions not declarations. This allows for removing the
dependency on FuncOp from Pass, and also more easily support non-FuncOp
function definition passes.

The definition of FunctionPass is left intact for now to allow time for downstream
users to migrate.

Differential Revision: https://reviews.llvm.org/D116632